# protocol/openapi.yaml
openapi: 3.1.0
info:
  title: CapOwn Control Plane API
  version: 1.6.0
  description: |
    Language-independent v1.6 wire contract for the CapOwn control plane.

    Covers the connectivity milestone, plugin discovery, task dispatch,
    claim-based job delivery, results, cancellation, and plugin invocation
    via the `plugin_call` task type. Workers claim jobs via long-poll;
    SSE is an optional wake accelerator.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  x-capown-protocol-version: "1.6"
  x-capown-plugin-error-codes:
    - plugin_not_found
    - plugin_disabled
    - plugin_unavailable
    - plugin_tool_not_found
    - plugin_schema_invalid
    - plugin_timeout
    - plugin_canceled
    - plugin_output_too_large
    - plugin_protocol_error
    - plugin_concurrency_exceeded
  x-capown-task-error-codes:
    - worker_offline
  x-capown-auth-error-codes:
    - user_disabled
    - invitation_invalid
servers:
  - url: http://localhost:9210
    description: Local development Master
...
paths:
...
  /v1/tokens:
    get:
      tags: [tokens]
      summary: List tokens owned by the current user
      operationId: listTokens
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [web]
      parameters:
        - $ref: '#/components/parameters/TokenTypeQuery'
      responses:
        "200":
          description: Token list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenListResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [tokens]
      summary: Create a client or admin token
      operationId: createToken
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [web]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTokenRequest'
      responses:
        "201":
          description: Token created. The plaintext token is returned only here.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedToken'
        "400":
          $ref: '#/components/responses/InvalidInput'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /v1/tokens/{token_id}:
    delete:
      tags: [tokens]
      summary: Revoke an owned token
      operationId: revokeToken
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [web]
      parameters:
        - $ref: '#/components/parameters/TokenId'
      responses:
        "204":
          description: Token revoked.
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Send `Authorization: Bearer <token>`. Token classes are identified by
        server-side storage and, for web/worker sessions, by the cown_* prefix.
  parameters:
...
    TokenId:
      name: token_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/ShortId'
...
    TokenTypeQuery:
      name: type
      in: query
      required: false
      schema:
        type: string
        enum: [client]
      description: When set to `client`, return only client tokens. Omit to list all owned tokens.
  responses:
    InvalidInput:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing, invalid, or expired credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Forbidden:
      description: Credentials are valid but do not grant the requested access.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/APIError'
    APIError:
      type: object
      additionalProperties: false
      required: [code, message, details]
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
          examples: [invalid_input]
        message:
          type: string
          description: Human-readable diagnostic message.
        details:
          description: Optional structured details. Current v1 commonly returns null.
          type: [object, array, string, number, boolean, 'null']
...
    CreateTokenRequest:
      type: object
      additionalProperties: false
      properties:
        type:
          type: string
          enum: [client, admin]
          default: client
        label:
          type: string
          default: ""
    CreatedToken:
      type: object
      additionalProperties: false
      required: [token_id, token, token_type, token_prefix, label, created_at]
      properties:
        token_id:
          $ref: '#/components/schemas/ShortId'
        token:
          type: string
          description: Plaintext token. Returned only at creation time.
        token_type:
          type: string
          enum: [client, admin]
        token_prefix:
          type: string
        label:
          type: string
        created_at:
          $ref: '#/components/schemas/Timestamp'
    TokenView:
      type: object
      additionalProperties: false
      required: [token_id, token_type, token_prefix, label, created_at, expires_at, last_used_at, revoked_at, status]
      properties:
        token_id:
          $ref: '#/components/schemas/ShortId'
        token_type:
          type: string
        token_prefix:
          type: string
        label:
          type: string
        created_at:
          $ref: '#/components/schemas/Timestamp'
        expires_at:
          type: [string, 'null']
          format: date-time
        last_used_at:
          type: [string, 'null']
          format: date-time
        revoked_at:
          type: [string, 'null']
          format: date-time
        status:
          type: string
          enum: [active, revoked]
    TokenListResponse:
      type: object
      additionalProperties: false
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TokenView'
        total:
          type: integer
          minimum: 0
...
    ShortId:
      type: string
      pattern: '^[0-9a-f]{12}$'
...
    Timestamp:
      type: string
      format: date-time
      description: UTC RFC 3339 timestamp.